+++ Platform.Sql/CustomerWallet.cs
@@ -16,7 +16,7 @@ namespace Platform.Sql
     {
         public int WalletId { get; set; }
         public int CustomerId { get; set; }
-        public long WalletBalance { get; set; }
+        public decimal WalletBalance { get; set; }
         public System.DateTime AmountDueDate { get; set; }
     }
 }
