+++ Platform.Sql/Product.cs
@@ -20,12 +20,12 @@ namespace Platform.Sql
             this.ProductSiteMappings = new HashSet<ProductSiteMapping>();
         }
     
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         public string ProductCode { get; set; }
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
-        public Nullable<int> ProductQuantity { get; set; }
-        public Nullable<long> ProductPrice { get; set; }
+        public Nullable<decimal> ProductQuantity { get; set; }
+        public Nullable<decimal> ProductPrice { get; set; }
         public Nullable<bool> IsActive { get; set; }
         public string Ref1 { get; set; }
         public string Ref2 { get; set; }
